apps/agents/requirements.txt
openai
ollama
python-dotenv
python-multipart
flake8
Pillow
beautifulsoup4
selenium
webdriver-manager
requests
pypdf2 
PyMuPDF
pdfminer.six
pdf2image
pytesseract

httpx
tenacity
structlog

pyspark

fastapi
uvicorn

# git+https://github.com/openai/swarm.git
